pyproject.toml
[tool.poetry]
name = "danpung-load-test"
version = "0.1.0"
description = ""
authors = ["jpark-olivawtech <jpark@olivawtech.com>"]
readme = "README.md"
packages = [{include = "danpung_load_test"}]

[tool.poetry.dependencies]
python = "^3.9"
lorem = "^0.1.1"
requests = "^2.28.1"
beautifulsoup4 = "^4.11.1"


[tool.poetry.group.dev.dependencies]
autopep8 = "^2.0.0"
pytest-asyncio = "^0.20.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
